Get rid of Events namespace, put everything inside Widget again
